fix(storefront): stop Add Google from hijacking accounts, guard the Zitadel finish route, and move session helpers out of use-server
