+++ apps/storefront/lib/auth/provider.test.ts
@@ -1,6 +1,10 @@
 import { afterEach, beforeEach, describe, expect, it } from "vitest";
 
-import { getAuthProvider, isGoogleSignInOffered } from "./provider";
+import {
+  getAuthProvider,
+  isGoogleLinkOffered,
+  isGoogleSignInOffered,
+} from "./provider";
 
 // Matches apps/storefront/app/sign-in/actions.ts's AUTH_PROVIDER rule
 // exactly: only the literal string "zitadel" flips the flag. This module
@@ -70,5 +74,38 @@ describe("isGoogleSignInOffered", () => {
   it('flag "true": Google sign-in is offered', () => {
     process.env.NEXT_PUBLIC_AUTH_PROVIDER = "true";
     expect(isGoogleSignInOffered()).toBe(true);
+  });
+});
+
+describe("isGoogleLinkOffered", () => {
+  // Unlike isGoogleSignInOffered, this one DOES gate on the flag:
+  // SecurityClient's "Add Google" (account-linking) control needs a
+  // "link this provider to my existing account" backend endpoint that
+  // does not exist under Zitadel — offering it there would let the
+  // control silently switch a signed-in shopper to a different,
+  // self-registered account. See the whole-branch security review's
+  // HIGH finding on this.
+  it('flag "zitadel": Google account-linking is NOT offered', () => {
+    process.env.NEXT_PUBLIC_AUTH_PROVIDER = "zitadel";
+    expect(isGoogleLinkOffered()).toBe(false);
+  });
+
+  it("flag unset: Google account-linking is offered (unchanged GIP link flow)", () => {
+    expect(isGoogleLinkOffered()).toBe(true);
+  });
+
+  it('flag "" (empty string): Google account-linking is offered', () => {
+    process.env.NEXT_PUBLIC_AUTH_PROVIDER = "";
+    expect(isGoogleLinkOffered()).toBe(true);
+  });
+
+  it('flag "Zitadel" (wrong case): Google account-linking is offered — exact literal match only', () => {
+    process.env.NEXT_PUBLIC_AUTH_PROVIDER = "Zitadel";
+    expect(isGoogleLinkOffered()).toBe(true);
+  });
+
+  it('flag "true": Google account-linking is offered', () => {
+    process.env.NEXT_PUBLIC_AUTH_PROVIDER = "true";
+    expect(isGoogleLinkOffered()).toBe(true);
   });
 });
