+++ apps/storefront/lib/auth/provider.ts
@@ -14,13 +14,28 @@ export function getAuthProvider(): "gip" | "zitadel" {
     : "gip";
 }
 
-// Google-through-Zitadel (phase 3c-2) now exists on both providers: under
-// GIP the control still goes through the mark8ly.com/auth/google
-// trampoline exactly as before, and under Zitadel it drives auth-bff's
-// own Google IDP intent instead (see @/lib/auth/google-sign-in). Kept as
-// its own function, rather than inlining `true` at each call site, so a
-// future gap in Google-through-Zitadel coverage (e.g. a third provider)
-// has one place to express "is Google offered at all" again.
+// Google-through-Zitadel (phase 3c-2) now exists on both providers for
+// SIGN-IN/SIGN-UP: under GIP the control still goes through the
+// mark8ly.com/auth/google trampoline exactly as before, and under
+// Zitadel it drives auth-bff's own Google IDP intent instead (see
+// @/lib/auth/google-sign-in). Kept as its own function, rather than
+// inlining `true` at each call site, so a future gap in
+// Google-through-Zitadel coverage (e.g. a third provider) has one place
+// to express "is Google offered at all" again.
 export function isGoogleSignInOffered(): boolean {
   return true;
+}
+
+// Whether SecurityClient's "Add Google" (account-linking) control may be
+// offered. This is deliberately NOT the same question as
+// isGoogleSignInOffered above: linking an existing, authenticated
+// shopper's account to a Google identity needs an authenticated
+// "link this provider to my existing account" backend endpoint, which
+// does not exist under Zitadel today — auth-bff's customer IDP endpoint
+// only self-registers/signs in. Offering the control under Zitadel would
+// let "Add Google" silently switch a signed-in shopper to a brand-new
+// account instead of linking one, so it must not be offered at all under
+// that flag. Under GIP the existing link flow is unaffected.
+export function isGoogleLinkOffered(): boolean {
+  return getAuthProvider() !== "zitadel";
 }
